[REFACTOR] Split supervisor.py into focused modules (958 -> 702 lines)
